+++ src/renderer/pages/Staking/index.ts
@@ -1,5 +1,3 @@
 export { Overview } from './Overview/Overview';
 export { Destination } from './Operations/Destination/Destination';
 export { Restake } from './Operations/Restake/Restake';
-export { StakeMore } from './Operations/StakeMore/StakeMore';
-export { ChangeValidators } from './Operations/ChangeValidators/ChangeValidators';
