+++ src/renderer/widgets/Staking/Withdraw/lib/types.ts
