+++ src/renderer/widgets/Staking/Withdraw/model/confirm-model.ts
