--- src/renderer/widgets/Unstake/index.ts
@@ -1,2 +1,0 @@
-export { Unstake } from './ui/Unstake';
-export { unstakeModel } from './model/unstake-model';
