--- src/renderer/widgets/Unstake/lib/unstake-utils.ts
@@ -1,29 +1,0 @@
-import { Step } from './types';
-
-export const unstakeUtils = {
-  isNoneStep,
-  isInitStep,
-  isConfirmStep,
-  isSignStep,
-  isSubmitStep,
-};
-
-function isNoneStep(step: Step): boolean {
-  return step === Step.NONE;
-}
-
-function isInitStep(step: Step): boolean {
-  return step === Step.INIT;
-}
-
-function isConfirmStep(step: Step): boolean {
-  return step === Step.CONFIRM;
-}
-
-function isSignStep(step: Step): boolean {
-  return step === Step.SIGN;
-}
-
-function isSubmitStep(step: Step): boolean {
-  return step === Step.SUBMIT;
-}
