--- src/renderer/widgets/Unstake/ui/UnstakeForm.tsx
@@ -1,302 +1,0 @@
-import { useForm } from 'effector-forms';
-import { FormEvent } from 'react';
-import { useUnit } from 'effector-react';
-
-import { useI18n } from '@app/providers';
-import { MultisigAccount } from '@shared/core';
-import { accountUtils, AccountAddress, ProxyWalletAlert } from '@entities/wallet';
-import { toAddress, toShortAddress, formatBalance } from '@shared/lib/utils';
-import { AssetBalance } from '@entities/asset';
-import { MultisigDepositWithLabel, FeeWithLabel } from '@entities/transaction';
-import { formModel } from '../model/form-model';
-import { Select, Input, Button, InputHint, AmountInput, MultiSelect, Shimmering } from '@shared/ui';
-import { DropdownOption } from '@shared/ui/types';
-
-type Props = {
-  onGoBack: () => void;
-};
-
-export const UnstakeForm = ({ onGoBack }: Props) => {
-  const { submit } = useForm(formModel.$unstakeForm);
-
-  const submitForm = (event: FormEvent) => {
-    event.preventDefault();
-    submit();
-  };
-
-  return (
-    <div className="pb-4 px-5">
-      <form id="transfer-form" className="flex flex-col gap-y-4 mt-4" onSubmit={submitForm}>
-        <ProxyFeeAlert />
-        <AccountsSelector />
-        <SignatorySelector />
-        <Amount />
-        <Description />
-      </form>
-      <div className="flex flex-col gap-y-6 pt-6 pb-4">
-        <FeeSection />
-      </div>
-      <ActionsSection onGoBack={onGoBack} />
-    </div>
-  );
-};
-
-const ProxyFeeAlert = () => {
-  const {
-    fields: { shards },
-  } = useForm(formModel.$unstakeForm);
-
-  const fee = useUnit(formModel.$fee);
-  const balance = useUnit(formModel.$proxyBalance);
-  const network = useUnit(formModel.$networkStore);
-  const proxyWallet = useUnit(formModel.$proxyWallet);
-
-  if (!network || !proxyWallet || !shards.hasError()) return null;
-
-  const formattedFee = formatBalance(fee, network.asset.precision).value;
-  const formattedBalance = formatBalance(balance, network.asset.precision).value;
-
-  return (
-    <ProxyWalletAlert
-      wallet={proxyWallet}
-      fee={formattedFee}
-      balance={formattedBalance}
-      symbol={network.asset.symbol}
-      onClose={shards.resetErrors}
-    />
-  );
-};
-
-const AccountsSelector = () => {
-  const { t } = useI18n();
-
-  const {
-    fields: { shards },
-  } = useForm(formModel.$unstakeForm);
-
-  const accounts = useUnit(formModel.$accounts);
-  const network = useUnit(formModel.$networkStore);
-
-  if (!network || accounts.length <= 1) return null;
-
-  const options = accounts.map(({ account, balances }) => {
-    const isShard = accountUtils.isShardAccount(account);
-    const address = toAddress(account.accountId, { prefix: network.chain.addressPrefix });
-
-    return {
-      id: account.id.toString(),
-      value: account,
-      element: (
-        <div className="flex justify-between w-full" key={account.id}>
-          <AccountAddress
-            size={20}
-            type="short"
-            address={address}
-            name={isShard ? toShortAddress(address, 16) : account.name}
-            canCopy={false}
-          />
-          <AssetBalance value={balances.stake} asset={network.asset} />
-        </div>
-      ),
-    };
-  });
-
-  return (
-    <div className="flex flex-col gap-y-2">
-      <MultiSelect
-        label={t('staking.bond.accountLabel')}
-        placeholder={t('staking.bond.accountPlaceholder')}
-        multiPlaceholder={t('staking.bond.manyAccountsPlaceholder')}
-        invalid={shards.hasError()}
-        selectedIds={shards.value.map((acc) => acc.id.toString())}
-        options={options}
-        onChange={(values) => shards.onChange(values.map(({ value }) => value))}
-      />
-      <InputHint variant="error" active={shards.hasError()}>
-        {t(shards.errorText())}
-      </InputHint>
-    </div>
-  );
-};
-
-const SignatorySelector = () => {
-  const { t } = useI18n();
-
-  const {
-    fields: { signatory },
-  } = useForm(formModel.$unstakeForm);
-
-  const signatories = useUnit(formModel.$signatories);
-  const isMultisig = useUnit(formModel.$isMultisig);
-  const network = useUnit(formModel.$networkStore);
-
-  if (!network || !isMultisig || signatories.length === 0) return null;
-
-  const options = signatories[0].reduce<DropdownOption[]>((acc, { signer, balance }) => {
-    if (!signer?.id) return acc;
-
-    const isShard = accountUtils.isShardAccount(signer);
-    const address = toAddress(signer.accountId, { prefix: network.chain.addressPrefix });
-
-    acc.push({
-      id: signer.id.toString(),
-      value: signer,
-      element: (
-        <div className="flex justify-between items-center w-full">
-          <AccountAddress
-            size={20}
-            type="short"
-            address={address}
-            name={isShard ? address : signer.name}
-            canCopy={false}
-          />
-          <AssetBalance value={balance} asset={network.asset} />
-        </div>
-      ),
-    });
-
-    return acc;
-  }, []);
-
-  return (
-    <div className="flex flex-col gap-y-2">
-      <Select
-        label={t('operation.selectSignatoryLabel')}
-        placeholder={t('operation.selectSignatory')}
-        selectedId={signatory.value.id?.toString()}
-        options={options}
-        invalid={signatory.hasError()}
-        onChange={({ value }) => signatory.onChange(value)}
-      />
-      <InputHint variant="error" active={signatory.hasError()}>
-        {t(signatory.errorText())}
-      </InputHint>
-    </div>
-  );
-};
-
-const Amount = () => {
-  const { t } = useI18n();
-
-  const {
-    fields: { amount },
-  } = useForm(formModel.$unstakeForm);
-
-  const unstakeBalanceRange = useUnit(formModel.$unstakeBalanceRange);
-  const isStakingLoading = useUnit(formModel.$isStakingLoading);
-  const network = useUnit(formModel.$networkStore);
-
-  if (!network) return null;
-
-  return (
-    <div className="flex flex-col gap-y-2">
-      <AmountInput
-        invalid={amount.hasError()}
-        value={amount.value}
-        balance={isStakingLoading ? <Shimmering width={50} height={10} /> : unstakeBalanceRange}
-        balancePlaceholder={t('general.input.availableLabel')}
-        placeholder={t('general.input.amountLabel')}
-        asset={network.asset}
-        onChange={amount.onChange}
-      />
-      <InputHint active={amount.hasError()} variant="error">
-        {t(amount.errorText())}
-      </InputHint>
-    </div>
-  );
-};
-
-const Description = () => {
-  const { t } = useI18n();
-
-  const {
-    fields: { description },
-  } = useForm(formModel.$unstakeForm);
-
-  const isMultisig = useUnit(formModel.$isMultisig);
-
-  if (!isMultisig) return null;
-
-  return (
-    <div className="flex flex-col gap-y-2">
-      <Input
-        spellCheck
-        className="w-full"
-        label={t('general.input.descriptionLabel')}
-        placeholder={t('general.input.descriptionPlaceholder')}
-        invalid={description.hasError()}
-        value={description.value}
-        onChange={description.onChange}
-      />
-      <InputHint active={description.hasError()} variant="error">
-        {t(description.errorText())}
-      </InputHint>
-    </div>
-  );
-};
-
-const FeeSection = () => {
-  const { t } = useI18n();
-
-  const {
-    fields: { shards },
-  } = useForm(formModel.$unstakeForm);
-
-  const api = useUnit(formModel.$api);
-  const network = useUnit(formModel.$networkStore);
-  const transactions = useUnit(formModel.$transactions);
-  const isMultisig = useUnit(formModel.$isMultisig);
-
-  if (!network || shards.value.length === 0) return null;
-
-  return (
-    <div className="flex flex-col gap-y-2">
-      {isMultisig && (
-        <MultisigDepositWithLabel
-          api={api}
-          asset={network.chain.assets[0]}
-          threshold={(shards.value[0] as MultisigAccount).threshold || 1}
-          onDepositChange={formModel.events.multisigDepositChanged}
-        />
-      )}
-
-      <FeeWithLabel
-        label={t('staking.networkFee', { count: shards.value.length || 1 })}
-        api={api}
-        asset={network.chain.assets[0]}
-        transaction={transactions?.[0]?.wrappedTx}
-        onFeeChange={formModel.events.feeChanged}
-        onFeeLoading={formModel.events.isFeeLoadingChanged}
-      />
-
-      {transactions && transactions.length > 1 && (
-        <FeeWithLabel
-          label={t('staking.networkFeeTotal')}
-          api={api}
-          asset={network.chain.assets[0]}
-          multiply={transactions.length}
-          transaction={transactions[0].wrappedTx}
-          onFeeChange={formModel.events.totalFeeChanged}
-          onFeeLoading={formModel.events.isFeeLoadingChanged}
-        />
-      )}
-    </div>
-  );
-};
-
-const ActionsSection = ({ onGoBack }: Props) => {
-  const { t } = useI18n();
-
-  const canSubmit = useUnit(formModel.$canSubmit);
-
-  return (
-    <div className="flex justify-between items-center mt-4">
-      <Button variant="text" onClick={onGoBack}>
-        {t('operation.goBackButton')}
-      </Button>
-      <Button form="transfer-form" type="submit" disabled={!canSubmit}>
-        {t('transfer.continueButton')}
-      </Button>
-    </div>
-  );
-};
