--- src/renderer/widgets/Withdraw/model/form-model.ts
@@ -1,657 +1,0 @@
-import { createEffect, attach, createEvent, createStore, combine, sample, restore, scopeBind } from 'effector';
-import { spread } from 'patronum';
-import { createForm } from 'effector-forms';
-import { BN } from '@polkadot/util';
-import { ApiPromise } from '@polkadot/api';
-import noop from 'lodash/noop';
-
-import { walletModel, walletUtils, accountUtils } from '@entities/wallet';
-import { balanceModel, balanceUtils } from '@entities/balance';
-import { networkModel, networkUtils } from '@entities/network';
-import type { Account, PartialBy, ProxiedAccount, Chain, Asset, Address, ChainId } from '@shared/core';
-import { useStakingData, StakingMap, useEra } from '@entities/staking';
-import {
-  transferableAmount,
-  getRelaychainAsset,
-  toAddress,
-  dictionary,
-  formatAmount,
-  ZERO_BALANCE,
-  redeemableAmount,
-} from '@shared/lib/utils';
-import { NetworkStore } from '../lib/types';
-import {
-  Transaction,
-  transactionBuilder,
-  transactionService,
-  MultisigTxWrapper,
-  ProxyTxWrapper,
-  DESCRIPTION_LENGTH,
-} from '@entities/transaction';
-
-type BalanceMap = { balance: string; redeemable: string };
-
-type FormParams = {
-  shards: Account[];
-  signatory: Account;
-  amount: string;
-  description: string;
-};
-
-type FormSubmitEvent = {
-  transactions: {
-    wrappedTx: Transaction;
-    multisigTx?: Transaction;
-    coreTx: Transaction;
-  }[];
-  formData: PartialBy<FormParams, 'signatory'> & {
-    proxiedAccount?: ProxiedAccount;
-    fee: string;
-    totalFee: string;
-    multisigDeposit: string;
-  };
-};
-
-const formInitiated = createEvent<NetworkStore>();
-const formSubmitted = createEvent<FormSubmitEvent>();
-const stakingSet = createEvent<StakingMap>();
-const eraSet = createEvent<number>();
-const formCleared = createEvent();
-
-const feeChanged = createEvent<string>();
-const totalFeeChanged = createEvent<string>();
-const multisigDepositChanged = createEvent<string>();
-const isFeeLoadingChanged = createEvent<boolean>();
-
-const $networkStore = createStore<{ chain: Chain; asset: Asset } | null>(null);
-const $staking = restore(stakingSet, null);
-const $era = restore(eraSet, null);
-const $minBond = createStore<string>(ZERO_BALANCE);
-const $stakingUnsub = createStore<() => void>(noop);
-const $eraUnsub = createStore<() => void>(noop);
-
-const $shards = createStore<Account[]>([]);
-const $isMultisig = createStore<boolean>(false);
-const $isProxy = createStore<boolean>(false);
-
-const $accountsBalances = createStore<BalanceMap[]>([]);
-const $redeemableBalance = createStore<string>(ZERO_BALANCE);
-const $signatoryBalance = createStore<string>(ZERO_BALANCE);
-const $proxyBalance = createStore<string>(ZERO_BALANCE);
-
-const $fee = restore(feeChanged, ZERO_BALANCE);
-const $totalFee = restore(totalFeeChanged, ZERO_BALANCE);
-const $multisigDeposit = restore(multisigDepositChanged, ZERO_BALANCE);
-const $isFeeLoading = restore(isFeeLoadingChanged, true);
-
-const $selectedSignatories = createStore<Account[]>([]);
-
-const $withdrawForm = createForm<FormParams>({
-  fields: {
-    shards: {
-      init: [] as Account[],
-      rules: [
-        {
-          name: 'noProxyFee',
-          source: combine({
-            fee: $fee,
-            isProxy: $isProxy,
-            proxyBalance: $proxyBalance,
-          }),
-          validator: (_s, _f, { isProxy, proxyBalance, fee }) => {
-            if (!isProxy) return true;
-
-            return new BN(fee).lte(new BN(proxyBalance));
-          },
-        },
-      ],
-    },
-    signatory: {
-      init: {} as Account,
-      rules: [
-        {
-          name: 'noSignatorySelected',
-          errorText: 'transfer.noSignatoryError',
-          source: $isMultisig,
-          validator: (signatory, _, isMultisig) => {
-            if (!isMultisig) return true;
-
-            return Object.keys(signatory).length > 0;
-          },
-        },
-        {
-          name: 'notEnoughTokens',
-          errorText: 'proxy.addProxy.notEnoughMultisigTokens',
-          source: combine({
-            fee: $fee,
-            isMultisig: $isMultisig,
-            multisigDeposit: $multisigDeposit,
-            signatoryBalance: $signatoryBalance,
-          }),
-          validator: (_s, _f, { fee, isMultisig, signatoryBalance, multisigDeposit }) => {
-            if (!isMultisig) return true;
-
-            return new BN(multisigDeposit).add(new BN(fee)).lte(new BN(signatoryBalance));
-          },
-        },
-      ],
-    },
-    amount: {
-      init: '',
-      rules: [
-        {
-          name: 'insufficientBalanceForFee',
-          errorText: 'transfer.notEnoughBalanceForFeeError',
-          source: combine({
-            network: $networkStore,
-            accountsBalances: $accountsBalances,
-          }),
-          validator: (value, form, { network, accountsBalances }) => {
-            const amountBN = new BN(formatAmount(value, network.asset.precision));
-
-            return form.shards.every((_: Account, index: number) => {
-              return amountBN.lte(new BN(accountsBalances[index].balance));
-            });
-          },
-        },
-      ],
-    },
-    description: {
-      init: '',
-      rules: [
-        {
-          name: 'maxLength',
-          errorText: 'transfer.descriptionLengthError',
-          validator: (value) => !value || value.length <= DESCRIPTION_LENGTH,
-        },
-      ],
-    },
-  },
-  validateOn: ['submit'],
-});
-
-// Effects
-type StakingParams = {
-  chainId: ChainId;
-  api: ApiPromise;
-  addresses: Address[];
-};
-const subscribeStakingFx = createEffect(({ chainId, api, addresses }: StakingParams): Promise<() => void> => {
-  const boundStakingSet = scopeBind(stakingSet, { safe: true });
-
-  return useStakingData().subscribeStaking(chainId, api, addresses, boundStakingSet);
-});
-
-const subscribeEraFx = createEffect((api: ApiPromise): Promise<() => void> => {
-  const boundEraSet = scopeBind(eraSet, { safe: true });
-
-  return useEra().subscribeActiveEra(api, (era) => {
-    if (!era) return;
-
-    boundEraSet(era);
-  });
-});
-
-// Computed
-
-const $txWrappers = combine(
-  {
-    wallet: walletModel.$activeWallet,
-    wallets: walletModel.$wallets,
-    shards: $shards,
-    accounts: walletModel.$accounts,
-    network: $networkStore,
-    signatories: $selectedSignatories,
-  },
-  ({ wallet, shards, accounts, wallets, network, signatories }) => {
-    if (!wallet || !network || shards.length !== 1) return [];
-
-    const walletFiltered = wallets.filter((wallet) => {
-      return !walletUtils.isProxied(wallet) && !walletUtils.isWatchOnly(wallet);
-    });
-    const walletsMap = dictionary(walletFiltered, 'id');
-    const chainFilteredAccounts = accounts.filter((account) => {
-      if (accountUtils.isBaseAccount(account) && walletUtils.isPolkadotVault(walletsMap[account.walletId])) {
-        return false;
-      }
-
-      return accountUtils.isChainAndCryptoMatch(account, network.chain);
-    });
-
-    return transactionService.getTxWrappers({
-      wallet,
-      wallets: walletFiltered,
-      account: shards[0],
-      accounts: chainFilteredAccounts,
-      signatories,
-    });
-  },
-);
-
-const $realAccounts = combine(
-  {
-    txWrappers: $txWrappers,
-    shards: $withdrawForm.fields.shards.$value,
-  },
-  ({ txWrappers, shards }) => {
-    if (shards.length === 0) return [];
-    if (txWrappers.length === 0) return shards;
-
-    if (transactionService.hasMultisig([txWrappers[0]])) {
-      return [(txWrappers[0] as MultisigTxWrapper).multisigAccount];
-    }
-
-    return [(txWrappers[0] as ProxyTxWrapper).proxyAccount];
-  },
-);
-
-const $proxyWallet = combine(
-  {
-    isProxy: $isProxy,
-    accounts: $realAccounts,
-    wallets: walletModel.$wallets,
-  },
-  ({ isProxy, accounts, wallets }) => {
-    if (!isProxy || accounts.length === 0) return undefined;
-
-    return walletUtils.getWalletById(wallets, accounts[0].walletId);
-  },
-  { skipVoid: false },
-);
-
-const $accounts = combine(
-  {
-    network: $networkStore,
-    wallet: walletModel.$activeWallet,
-    shards: $shards,
-    era: $era,
-    staking: $staking,
-    balances: balanceModel.$balances,
-  },
-  ({ network, wallet, era, shards, staking, balances }) => {
-    if (!wallet || !network || !staking) return [];
-
-    const { chain, asset } = network;
-
-    return shards.map((shard) => {
-      const balance = balanceUtils.getBalance(balances, shard.accountId, chain.chainId, asset.assetId.toString());
-      const address = toAddress(shard.accountId, { prefix: chain.addressPrefix });
-      const redeemable = redeemableAmount(staking[address]?.unlocking, era || 0);
-
-      return {
-        account: shard,
-        balances: { balance: transferableAmount(balance), redeemable },
-      };
-    });
-  },
-);
-
-const $signatories = combine(
-  {
-    network: $networkStore,
-    txWrappers: $txWrappers,
-    balances: balanceModel.$balances,
-  },
-  ({ network, txWrappers, balances }) => {
-    if (!network) return [];
-
-    const { chain, asset } = network;
-
-    return txWrappers.reduce<Array<{ signer: Account; balance: string }[]>>((acc, wrapper) => {
-      if (!transactionService.hasMultisig([wrapper])) return acc;
-
-      const balancedSignatories = (wrapper as MultisigTxWrapper).signatories.map((signatory) => {
-        const balance = balanceUtils.getBalance(balances, signatory.accountId, chain.chainId, asset.assetId.toString());
-
-        return { signer: signatory, balance: transferableAmount(balance) };
-      });
-
-      acc.push(balancedSignatories);
-
-      return acc;
-    }, []);
-  },
-);
-
-const $isChainConnected = combine(
-  {
-    network: $networkStore,
-    statuses: networkModel.$connectionStatuses,
-  },
-  ({ network, statuses }) => {
-    if (!network) return false;
-
-    return networkUtils.isConnectedStatus(statuses[network.chain.chainId]);
-  },
-);
-
-const $api = combine(
-  {
-    apis: networkModel.$apis,
-    network: $networkStore,
-  },
-  ({ apis, network }) => {
-    if (!network) return undefined;
-
-    return apis[network.chain.chainId];
-  },
-  { skipVoid: false },
-);
-
-const $pureTxs = combine(
-  {
-    network: $networkStore,
-    form: $withdrawForm.$values,
-    isConnected: $isChainConnected,
-  },
-  ({ network, form, isConnected }) => {
-    if (!network || !isConnected) return undefined;
-
-    return form.shards.map((shard) => {
-      return transactionBuilder.buildRedeem({
-        chain: network.chain,
-        accountId: shard.accountId,
-      });
-    });
-  },
-  { skipVoid: false },
-);
-
-const $transactions = combine(
-  {
-    apis: networkModel.$apis,
-    networkStore: $networkStore,
-    pureTxs: $pureTxs,
-    txWrappers: $txWrappers,
-  },
-  ({ apis, networkStore, pureTxs, txWrappers }) => {
-    if (!networkStore || !pureTxs) return undefined;
-
-    return pureTxs.map((tx) =>
-      transactionService.getWrappedTransaction({
-        api: apis[networkStore.chain.chainId],
-        addressPrefix: networkStore.chain.addressPrefix,
-        transaction: tx,
-        txWrappers,
-      }),
-    );
-  },
-  { skipVoid: false },
-);
-
-const $canSubmit = combine(
-  {
-    isFormValid: $withdrawForm.$isValid,
-    isFeeLoading: $isFeeLoading,
-    isStakingLoading: subscribeStakingFx.pending,
-    isEraLoading: subscribeEraFx.pending,
-  },
-  ({ isFormValid, isFeeLoading, isStakingLoading, isEraLoading }) => {
-    return isFormValid && !isFeeLoading && !isStakingLoading && !isEraLoading;
-  },
-);
-
-// Fields connections
-
-sample({
-  clock: formInitiated,
-  target: [$withdrawForm.reset, $selectedSignatories.reinit],
-});
-
-sample({
-  clock: formInitiated,
-  filter: ({ chain, shards }) => Boolean(getRelaychainAsset(chain.assets)) && shards.length > 0,
-  fn: ({ chain, shards }) => ({
-    shards,
-    networkStore: { chain, asset: getRelaychainAsset(chain.assets)! },
-  }),
-  target: spread({
-    shards: $shards,
-    networkStore: $networkStore,
-  }),
-});
-
-sample({
-  clock: formInitiated,
-  source: {
-    networkStore: $networkStore,
-    api: $api,
-    shards: $shards,
-  },
-  filter: ({ networkStore, api }) => {
-    return Boolean(networkStore) && Boolean(api);
-  },
-  fn: ({ networkStore, api, shards }) => {
-    const addresses = shards.map((shard) => toAddress(shard.accountId, { prefix: networkStore!.chain.addressPrefix }));
-
-    return {
-      chainId: networkStore!.chain.chainId,
-      api: api!,
-      addresses,
-    };
-  },
-  target: subscribeStakingFx,
-});
-
-sample({
-  clock: formInitiated,
-  source: $api,
-  filter: (api): api is ApiPromise => Boolean(api),
-  target: subscribeEraFx,
-});
-
-sample({
-  clock: subscribeStakingFx.doneData,
-  target: $stakingUnsub,
-});
-
-sample({
-  clock: subscribeEraFx.doneData,
-  target: $eraUnsub,
-});
-
-sample({
-  clock: $accountsBalances,
-  filter: (balances) => Boolean(balances),
-  fn: (balances) => {
-    if (balances.length === 0) return ZERO_BALANCE;
-
-    const totalRedeemable = balances.reduce<BN>((acc, { redeemable }) => {
-      if (!redeemable) return acc;
-
-      return new BN(redeemable).add(new BN(acc));
-    }, new BN(ZERO_BALANCE));
-
-    return totalRedeemable.toString();
-  },
-  target: $redeemableBalance,
-});
-
-sample({
-  clock: formInitiated,
-  source: $shards,
-  filter: (shards) => shards.length > 0,
-  target: $withdrawForm.fields.shards.onChange,
-});
-
-sample({
-  source: {
-    accounts: $accounts,
-    shards: $withdrawForm.fields.shards.$value,
-  },
-  fn: ({ accounts, shards }) => {
-    return accounts.reduce<{ balance: string; redeemable: string }[]>((acc, { account, balances }) => {
-      if (shards.includes(account)) {
-        acc.push(balances);
-      }
-
-      return acc;
-    }, []);
-  },
-  target: $accountsBalances,
-});
-
-sample({
-  clock: $withdrawForm.fields.signatory.onChange,
-  source: $signatories,
-  filter: (signatories) => signatories.length > 0,
-  fn: (signatories, signatory) => {
-    const match = signatories[0].find(({ signer }) => signer.id === signatory.id);
-
-    return match?.balance || ZERO_BALANCE;
-  },
-  target: $signatoryBalance,
-});
-
-sample({
-  clock: $withdrawForm.fields.signatory.$value,
-  fn: (signatory) => [signatory],
-  target: $selectedSignatories,
-});
-
-sample({
-  clock: $withdrawForm.fields.shards.onChange,
-  target: $withdrawForm.fields.amount.resetErrors,
-});
-
-sample({
-  clock: $withdrawForm.fields.amount.onChange,
-  target: $withdrawForm.fields.shards.resetErrors,
-});
-
-sample({
-  clock: $txWrappers.updates,
-  fn: (txWrappers) => ({
-    isProxy: transactionService.hasProxy(txWrappers),
-    isMultisig: transactionService.hasMultisig(txWrappers),
-  }),
-  target: spread({
-    isProxy: $isProxy,
-    isMultisig: $isMultisig,
-  }),
-});
-
-sample({
-  clock: $realAccounts.updates,
-  source: {
-    isProxy: $isProxy,
-    balances: balanceModel.$balances,
-    network: $networkStore,
-  },
-  filter: ({ isProxy, network }, accounts) => {
-    return isProxy && Boolean(network) && accounts.length > 0;
-  },
-  fn: ({ balances, network }, accounts) => {
-    const balance = balanceUtils.getBalance(
-      balances,
-      accounts[0].accountId,
-      network!.chain.chainId,
-      network!.asset.assetId.toString(),
-    );
-
-    return transferableAmount(balance);
-  },
-  target: $proxyBalance,
-});
-
-// Submit
-
-sample({
-  clock: $withdrawForm.formValidated,
-  source: {
-    amount: $redeemableBalance,
-    realAccounts: $realAccounts,
-    network: $networkStore,
-    transactions: $transactions,
-    isProxy: $isProxy,
-    fee: $fee,
-    totalFee: $totalFee,
-    multisigDeposit: $multisigDeposit,
-  },
-  filter: ({ network, transactions }) => {
-    return Boolean(network) && Boolean(transactions);
-  },
-  fn: ({ amount, realAccounts, network, transactions, isProxy, ...fee }, formData) => {
-    const { shards, ...rest } = formData;
-
-    const signatory = formData.signatory.accountId ? formData.signatory : undefined;
-    // TODO: update after i18n effector integration
-    const defaultText = `Redeem ${formData.amount} ${network!.asset.symbol}`;
-    const description = signatory ? formData.description || defaultText : '';
-
-    return {
-      transactions: transactions!.map((tx) => ({
-        wrappedTx: tx.wrappedTx,
-        multisigTx: tx.multisigTx,
-        coreTx: tx.coreTx,
-      })),
-      formData: {
-        ...fee,
-        ...rest,
-        shards: realAccounts,
-        amount,
-        signatory,
-        description,
-        ...(isProxy && { proxiedAccount: shards[0] as ProxiedAccount }),
-      },
-    };
-  },
-  target: formSubmitted,
-});
-
-sample({
-  clock: formSubmitted,
-  target: attach({
-    source: $stakingUnsub,
-    effect: (unsub) => unsub(),
-  }),
-});
-
-sample({
-  clock: formSubmitted,
-  target: attach({
-    source: $eraUnsub,
-    effect: (unsub) => unsub(),
-  }),
-});
-
-sample({
-  clock: formCleared,
-  target: [$withdrawForm.reset, $shards.reinit],
-});
-
-export const formModel = {
-  $withdrawForm,
-  $proxyWallet,
-  $signatories,
-
-  $accounts,
-  $accountsBalances,
-  $redeemableBalance,
-  $proxyBalance,
-
-  $fee,
-  $multisigDeposit,
-
-  $api,
-  $networkStore,
-  $transactions,
-  $isMultisig,
-  $isChainConnected,
-  $isStakingLoading: subscribeStakingFx.pending,
-  $isEraLoading: subscribeEraFx.pending,
-  $canSubmit,
-
-  events: {
-    formInitiated,
-    formCleared,
-
-    feeChanged,
-    totalFeeChanged,
-    multisigDepositChanged,
-    isFeeLoadingChanged,
-  },
-  output: {
-    formSubmitted,
-  },
-};
